requirements.txt
GitPython
python-dotenv
langchain
langchain-google-genai
Flask
gunicorn # For production deployment in Docker